+++ collect.go
@@ -23,9 +23,9 @@ type CollectServiceOp struct {
 
 // Collect represents a Shopify collect
 type Collect struct {
-	ID           int        `json:"id,omitempty"`
-	CollectionID int        `json:"collection_id,omitempty"`
-	ProductID    int        `json:"product_id,omitempty"`
+	ID           int64      `json:"id,omitempty"`
+	CollectionID int64      `json:"collection_id,omitempty"`
+	ProductID    int64      `json:"product_id,omitempty"`
 	Featured     bool       `json:"featured,omitempty"`
 	CreatedAt    *time.Time `json:"created_at,omitempty"`
 	UpdatedAt    *time.Time `json:"updated_at,omitempty"`
