+++ customcollection.go
@@ -14,10 +14,10 @@ const customCollectionsResourceName = "collections"
 type CustomCollectionService interface {
 	List(interface{}) ([]CustomCollection, error)
 	Count(interface{}) (int, error)
-	Get(int, interface{}) (*CustomCollection, error)
+	Get(int64, interface{}) (*CustomCollection, error)
 	Create(CustomCollection) (*CustomCollection, error)
 	Update(CustomCollection) (*CustomCollection, error)
-	Delete(int) error
+	Delete(int64) error
 
 	// MetafieldsService used for CustomCollection resource to communicate with Metafields resource
 	MetafieldsService
@@ -31,7 +31,7 @@ type CustomCollectionServiceOp struct {
 
 // CustomCollection represents a Shopify custom collection.
 type CustomCollection struct {
-	ID             int         `json:"id"`
+	ID             int64       `json:"id"`
 	Handle         string      `json:"handle"`
 	Title          string      `json:"title"`
 	UpdatedAt      *time.Time  `json:"updated_at"`
@@ -70,7 +70,7 @@ func (s *CustomCollectionServiceOp) Count(options interface{}) (int, error) {
 }
 
 // Get individual custom collection
-func (s *CustomCollectionServiceOp) Get(collectionID int, options interface{}) (*CustomCollection, error) {
+func (s *CustomCollectionServiceOp) Get(collectionID int64, options interface{}) (*CustomCollection, error) {
 	path := fmt.Sprintf("%s/%d.json", customCollectionsBasePath, collectionID)
 	resource := new(CustomCollectionResource)
 	err := s.client.Get(path, resource, options)
@@ -97,42 +97,42 @@ func (s *CustomCollectionServiceOp) Update(collection CustomCollection) (*Custom
 }
 
 // Delete an existing custom collection.
-func (s *CustomCollectionServiceOp) Delete(collectionID int) error {
+func (s *CustomCollectionServiceOp) Delete(collectionID int64) error {
 	return s.client.Delete(fmt.Sprintf("%s/%d.json", customCollectionsBasePath, collectionID))
 }
 
 // List metafields for a custom collection
-func (s *CustomCollectionServiceOp) ListMetafields(customCollectionID int, options interface{}) ([]Metafield, error) {
+func (s *CustomCollectionServiceOp) ListMetafields(customCollectionID int64, options interface{}) ([]Metafield, error) {
 	metafieldService := &MetafieldServiceOp{client: s.client, resource: customCollectionsResourceName, resourceID: customCollectionID}
 	return metafieldService.List(options)
 }
 
 // Count metafields for a custom collection
-func (s *CustomCollectionServiceOp) CountMetafields(customCollectionID int, options interface{}) (int, error) {
+func (s *CustomCollectionServiceOp) CountMetafields(customCollectionID int64, options interface{}) (int, error) {
 	metafieldService := &MetafieldServiceOp{client: s.client, resource: customCollectionsResourceName, resourceID: customCollectionID}
 	return metafieldService.Count(options)
 }
 
 // Get individual metafield for a custom collection
-func (s *CustomCollectionServiceOp) GetMetafield(customCollectionID int, metafieldID int, options interface{}) (*Metafield, error) {
+func (s *CustomCollectionServiceOp) GetMetafield(customCollectionID int64, metafieldID int64, options interface{}) (*Metafield, error) {
 	metafieldService := &MetafieldServiceOp{client: s.client, resource: customCollectionsResourceName, resourceID: customCollectionID}
 	return metafieldService.Get(metafieldID, options)
 }
 
 // Create a new metafield for a custom collection
-func (s *CustomCollectionServiceOp) CreateMetafield(customCollectionID int, metafield Metafield) (*Metafield, error) {
+func (s *CustomCollectionServiceOp) CreateMetafield(customCollectionID int64, metafield Metafield) (*Metafield, error) {
 	metafieldService := &MetafieldServiceOp{client: s.client, resource: customCollectionsResourceName, resourceID: customCollectionID}
 	return metafieldService.Create(metafield)
 }
 
 // Update an existing metafield for a custom collection
-func (s *CustomCollectionServiceOp) UpdateMetafield(customCollectionID int, metafield Metafield) (*Metafield, error) {
+func (s *CustomCollectionServiceOp) UpdateMetafield(customCollectionID int64, metafield Metafield) (*Metafield, error) {
 	metafieldService := &MetafieldServiceOp{client: s.client, resource: customCollectionsResourceName, resourceID: customCollectionID}
 	return metafieldService.Update(metafield)
 }
 
 // // Delete an existing metafield for a custom collection
-func (s *CustomCollectionServiceOp) DeleteMetafield(customCollectionID int, metafieldID int) error {
+func (s *CustomCollectionServiceOp) DeleteMetafield(customCollectionID int64, metafieldID int64) error {
 	metafieldService := &MetafieldServiceOp{client: s.client, resource: customCollectionsResourceName, resourceID: customCollectionID}
 	return metafieldService.Delete(metafieldID)
 }
